LayoutTests/platform/ios-simulator/editing/pasteboard/paste-and-sanitize-expected.txt

This test checks that the paste operation trims the pasted fragment to reduce the verbosity of the markup without affecting the style.

FAIL confirmedMarkup should be Hello. Was <span style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">Hello</span>.
FAIL confirmedMarkup should be <b><i>Hello</i></b>. Was <b><i style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">Hello</i></b>.
FAIL confirmedMarkup should be <b><i>Hello</i></b>. Was <b><i style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">Hello</i></b>.
FAIL confirmedMarkup should be Hello. Was <span style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">Hello</span>.
FAIL confirmedMarkup should be <i style="font-weight: bold;">Hello</i>. Was <i style="font-weight: bold; -webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">Hello</i>.
FAIL confirmedMarkup should be <div style="text-align: center;"><b>Hello</b></div>. Was <div style="text-align: start;"><b style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">Hello</b></div>.
FAIL confirmedMarkup should be <div><b><i>hello</i></b></div><div><b><i>world</i></b></div>. Was <div><b><i style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">hello</i></b></div><div><b><i style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">world</i></b></div>.
FAIL confirmedMarkup should be <b><i><span style="font-weight: normal;"><b><i>hello1</i></b><b><i>&nbsp;hello2</i></b></span></i></b>. Was <b><i><span style="font-weight: normal; -webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"><b><i>hello1</i></b><b><i>&nbsp;hello2</i></b></span></i></b>.
FAIL confirmedMarkup should be <i style="margin: 10px;">hello</i></b></i>. Was <i style="margin: 10px;"><b><i style="margin: 10px; -webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">hello</i></b></i>.
FAIL confirmedMarkup should be <b><i>Hello&nbsp;world</i></b>. Was <b><i style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">Hello&nbsp;world</i></b>.
FAIL confirmedMarkup should be <b><i><span style="font-weight: normal;">plain text<b><i>bold italic text</i></b></span></i></b>. Was <b><i><span style="font-weight: normal; -webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);">plain text<b><i>bold italic text</i></b></span></i></b>.
PASS successfullyParsed is true

TEST COMPLETE
